requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
openpyxl>=3.1.0
Pillow>=9.0.0
firebase-admin>=6.0.0

